backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase==2.9.1
google-api-python-client==2.151.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
